made Report compatible to ResponsiveUIValidator. remove first version of compatible ResponsiveUIValidator (to be replace by Report).
